Fixed upper-case machine-name bug introduced into run_acme
It appears that the code to make $machine lowercase got moved from
before setting up the batch queues, to afterwards.   This means that
the batch settings may not be set properly if the machine name contains
capital letters.

The bug was probably introduced on May 15 or 17.

This commit fixes the bug.

This commit also moves some comments around in the code so they are
in more appropriate locations.