Remove support for CLM4.0
CLM4.0 is being dropped in CESM, and has already been dropped in
E3SM. This commit removes some of the complexity related to supporting
that old version of CLM.

This commit completely removes support for CLM4.0 for CESM. For E3SM,
support is removed in files that are shared between CESM and E3SM, but I
have not made any changes to E3SM-specific files.

A key change was in the build: we no longer need to distinguish between
CLM4.0 vs. later versions of CLM in setting some build options (but I
think that we *do* still need to determine whether we're running with
CLM or some other land model, including dlnd/slnd/xlnd, so some
CLM-specific logic still remains in the Makefile).

This change is backwards-incompatible with respect to CLM4.0 support,
but should be backwards compatible for cases using later versions of
CLM, as well as cases using other land components.

Fixes ESMCI/cime#2963