Point to pFUnit v4 builds on cheyenne, casper and izumi
We will no longer support pFUnit v3 (there are backwards compatibility
issues with trying to support both v3 and v4).

Note that I have removed support for an MPI-enabled build of pFUnit on
cheyenne: we weren't using the MPI capabilities of pFUnit, and I suspect
they currently won't work (we would probably need some updates in CIME's
CMake and/or run_tests infrastructure to allow this to work correctly).