Use gfortran v10+ behavior for undefined CMAKE_Fortran_COMPILER_VERSION
The typical CESM unit test CMake-based builds include
CIME_initial_setup.cmake before the `project` line (as is recommended in
comments in CIME_initial_setup.cmake). This pulls in the Macros at that
point. However, CMAKE_Fortran_COMPILER_VERSION is not defined until
after the `project` line. This causes problems on machines that need
this logic for gfortran 10+. I have worked around it before by defining
FFLAGS unconditionally in the machine-specific cmake file (e.g.,
gnu_green.cmake), but now that most machines are using gfortran 10+, it
seems like the cleaner fix is to use the gfortran 10+ behavior as the
default if CMAKE_Fortran_COMPILER_VERSION is not yet defined.